Add find by name, delete, inspect, and load data methods to repo module
